+++ src/slave/configuration.rs
@@ -296,7 +296,7 @@ impl<'a> SlaveRef<'a, &'a mut Slave> {
 
         // ETG1000.6 Table 67 – CoE Communication Area
         let num_sms = self
-            .read_sdo::<u8>(SM_TYPE_ADDRESS, SubIndex::Index(0))
+            .sdo_read::<u8>(SM_TYPE_ADDRESS, SubIndex::Index(0))
             .await?;
 
         log::trace!("Found {num_sms} SMs from CoE");
@@ -312,7 +312,7 @@ impl<'a> SlaveRef<'a, &'a mut Slave> {
         // NOTE: This is a 1-based SDO sub-index
         for sm_mapping_sub_index in sm_range {
             let sm_type = self
-                .read_sdo::<u8>(SM_TYPE_ADDRESS, SubIndex::Index(sm_mapping_sub_index))
+                .sdo_read::<u8>(SM_TYPE_ADDRESS, SubIndex::Index(sm_mapping_sub_index))
                 .await
                 .map(SyncManagerType::from_primitive)?;
 
@@ -333,20 +333,20 @@ impl<'a> SlaveRef<'a, &'a mut Slave> {
             }
 
             // Total number of PDO assignments for this sync manager
-            let num_sm_assignments = self.read_sdo::<u8>(sm_address, SubIndex::Index(0)).await?;
+            let num_sm_assignments = self.sdo_read::<u8>(sm_address, SubIndex::Index(0)).await?;
 
             log::trace!("SDO sync manager {sync_manager_index} (sub index #{sm_mapping_sub_index}) {sm_address:#06x} {sm_type:?}, sub indices: {num_sm_assignments}");
 
             let mut sm_bit_len = 0u16;
 
             for i in 1..=num_sm_assignments {
-                let pdo = self.read_sdo::<u16>(sm_address, SubIndex::Index(i)).await?;
-                let num_mappings = self.read_sdo::<u8>(pdo, SubIndex::Index(0)).await?;
+                let pdo = self.sdo_read::<u16>(sm_address, SubIndex::Index(i)).await?;
+                let num_mappings = self.sdo_read::<u8>(pdo, SubIndex::Index(0)).await?;
 
                 log::trace!("--> #{i} data: {pdo:#06x} ({num_mappings} mappings):");
 
                 for i in 1..=num_mappings {
-                    let mapping = self.read_sdo::<u32>(pdo, SubIndex::Index(i)).await?;
+                    let mapping = self.sdo_read::<u32>(pdo, SubIndex::Index(i)).await?;
 
                     // Yes, big-endian. Makes life easier when mapping from debug prints to actual
                     // data fields.